fix: move errors to error file
