+++ request/parser.go
@@ -14,13 +14,6 @@ import (
 	"github.com/buildwithgrove/path/relayer"
 )
 
-const (
-	// TODO_DISCUSS: should we create a common errors package for all errors?
-	// TODO_IMPROVE: define specific error codes for all errors
-	parserErrorCode     = -32000
-	parserErrorTemplate = `{"jsonrpc":"2.0","id":"0","error":{"code":%d,"message":"%s"}}`
-)
-
 type (
 	Parser struct {
 		Backend            Backend
@@ -68,6 +61,7 @@ func (p *Parser) GetQoSService(ctx context.Context, req *http.Request) (relayer.
 }
 
 func (p *Parser) GetHTTPErrorResponse(ctx context.Context, err error) gateway.HTTPResponse {
+	// TODO_TECHDEBT: return the correct id field from the request if the request is a valid JSON-RPC
 	return &ParserErrorResponse{err: err.Error()}
 }
 
@@ -89,22 +83,4 @@ func (p *Parser) getServiceID(host string) (relayer.ServiceID, error) {
 	}
 
 	return serviceID, nil
-}
-
-/* Parser Error Response */
-
-type ParserErrorResponse struct {
-	err string
-}
-
-func (r *ParserErrorResponse) GetPayload() []byte {
-	return []byte(fmt.Sprintf(parserErrorTemplate, parserErrorCode, r.err))
-}
-
-func (r *ParserErrorResponse) GetHTTPStatusCode() int {
-	return http.StatusBadRequest
-}
-
-func (r *ParserErrorResponse) GetHTTPHeaders() map[string]string {
-	return map[string]string{}
 }
